refactor auth
move files
add loginButton and logoutButton
add login page
