+++ app/routes/auth.callback.tsx
@@ -1,5 +1,5 @@
 import { redirect, type LoaderFunctionArgs } from '@remix-run/node'
-import { createServerClient } from '~/utils/supabase.server'
+import { createServerClient } from '~/lib/supabase.server'
 
 export async function loader({ request }: LoaderFunctionArgs) {
   const requestUrl = new URL(request.url)
