+++ app/routes/dashboard.tsx
@@ -1,18 +1,18 @@
 import { LoaderFunction, json } from '@remix-run/node'
-import Login from '~/components/login'
-import { requireSession } from '~/utils/session.server'
+import LogoutButton from '~/components/logoutButton'
+import { requireUser } from '~/lib/session.server'
 
 export const loader: LoaderFunction = async ({ request }) => {
-  const session = await requireSession(request)
+  const user = await requireUser(request)
 
-  return json({ session })
+  return json({ user })
 }
 
 export default function Dashboard() {
   return (
     <div>
       <h1>Dashboard</h1>
-      <Login />
+      <LogoutButton />
     </div>
   )
 }
